requirements.txt
pytest==8.3.5
requests==2.32.3
allure-pytest==2.14.2
allure-python-commons==2.14.2
playwright==1.52.0
pytest-playwright==0.7.0
ruff==0.11.10
